--- menu/menu.cpp
@@ -1,16 +1,0 @@
-#include <iostream>
-#include "../menu_header/menu.h"
-#include "../menu_header/mecanismo_menu.h"
-
-using namespace std;
-
-// solo muestra texto del menu
-void mostrarMenu() {
-
-    cout << " 1 - Jugar" << endl;
-	cout << " 2 - Estadistica" << endl;
-	cout << " 3 - Creditos" << endl;
-	cout << " 4 - Reglas" << endl;
-	cout << " ------------------ " << endl;
-	cout << " 0 - Salir" << endl;
-}
