--- menu/reglas.cpp
@@ -1,14 +1,0 @@
-#include <iostream>
-#include "../menu_header/reglas.h"
-#include "../menu_header/mecanismo_menu.h"
-
-using namespace std;
-
-void mostrarReglas() {
-
-    cout << "agregar las reglas masomenos" << endl;
-
-    cout << endl << "2 - volver";
-
-    volver();
-}
