--- menu_header/creditos.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void mostrarCreditos();
