--- menu_header/mecanismo_menu.h
@@ -1,7 +1,0 @@
-#pragma once
-
-void mecanismoMenu();
-
-void volver();
-
-void espaciosBlancos();
