--- menu_header/menu.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void mostrarMenu();
