--- menu_header/reglas.h
@@ -1,6 +1,0 @@
-#ifndef REGLAS_H_INCLUDED
-#define REGLAS_H_INCLUDED
-
-void mostrarReglas();
-
-#endif // REGLAS_H_INCLUDED
