+++ menus.cpp
@@ -1,12 +1,62 @@
 #include <iostream>
-#include <conio.h>
-#include "../menu_header/mecanismo_menu.h"
-#include "../menu_header/creditos.h"
-#include "../menu_header/estadisticas.h"
-#include "../menu_header/menu.h"
-#include "../juego_header/suma_dados.h"
-#include "../juego_header/juego.h"
-#include "../menu_header/reglas.h"
+#include "menus.h"
+#include "conio.h"
+#include "jueguito.h"
+
+using namespace std;
+
+// solo muestra texto del menu
+void mostrarMenu() {
+
+    cout << " 1 - Jugar" << endl;
+	cout << " 2 - Estadistica" << endl;
+	cout << " 3 - Creditos" << endl;
+	cout << " 4 - Reglas" << endl;
+	cout << " ------------------ " << endl;
+	cout << " 0 - Salir" << endl;
+}
+
+void mostrarCreditos() {
+
+    cout<<"=============================" << endl;
+    cout<< "          Creditos           " << endl;
+    cout<<"============================="<< endl << endl;
+
+    cout<< "  Materia: Progamacion 1"<<endl;
+    cout<< "  Profesor: Angel Simon"<< endl << endl;
+
+    cout<< "  Integrantes de 'Los Pibes'"<<endl;
+    cout<< "  1. Tomas Centurion     - 30926"<<endl;
+    cout<< "  2. Franco Medero       - 27589"<< endl;
+    cout<< "  3. Sebastian Gomez     - 28781"<< endl;
+    cout<< "  4. Angel Diaz          - 33472"<< endl;
+
+
+    cout << endl << endl << "2 - volver" << endl;
+
+    volver();
+}
+
+// solo muestra el texto de estadisticas
+void mostarEstadisticas (vector<int>& puntosEstadisticas, vector<string>& nombresEstadisticas) {
+    cout << "---- ESTADISTICAS ----" << endl;
+    for(int i=0; i<nombresEstadisticas.size(); i++) {
+        cout << "# " << nombresEstadisticas[i] << " - " << puntosEstadisticas[i] << "pts" << endl;
+    }
+
+    cout << endl << endl << "2 - volver" << endl;
+
+    volver();
+}
+
+void mostrarReglas() {
+
+    cout << "agregar las reglas masomenos" << endl;
+
+    cout << endl << "2 - volver";
+
+    volver();
+}
 
 // es el mecanismo no toquen nada jeje
 void mecanismoMenu() {
@@ -53,6 +103,7 @@ void espaciosBlancos() {
     for (int i = 0; i < 50; ++i) {
         std::cout << "\n";
     }
+
 }
 
 // boton de volver, si se elige 2 se vuelve al
