+++ menus.h
@@ -1,6 +1,20 @@
 #pragma once
+
 #include <iostream>
 #include <vector>
+
 using namespace std;
 
+void mostrarCreditos();
+
 void mostarEstadisticas (vector<int>& puntosEstadisticas, vector<string>& nombresEstadisticas);
+
+void mostrarMenu();
+
+void mecanismoMenu();
+
+void volver();
+
+void espaciosBlancos();
+
+void mostrarReglas();
